requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
beautifulsoup4
requests
sentence-transformers
chromadb
pypdf
python-docx
langchain-text-splitters
google-genai
